moved memory test
